fix: create preference

+++ management/src/main/java/team/proximity/management/services/PreferenceService.java
@@ -1,5 +1,9 @@
 package team.proximity.management.services;
 
+import team.proximity.management.dto.BookingDayDTO;
+import team.proximity.management.dto.PreferenceDTO;
+import team.proximity.management.exceptions.PreferenceNotFoundException;
+import team.proximity.management.model.BookingDay;
 import team.proximity.management.model.Preference;
 import org.springframework.stereotype.Service;
 import team.proximity.management.repositories.PreferenceRepository;
@@ -8,6 +12,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class PreferenceService {
@@ -17,32 +22,50 @@ public class PreferenceService {
         this.preferenceRepository = preferenceRepository;
     }
 
-    public Preference createPreference(Preference preference) {
+    public Preference createPreference(PreferenceDTO preferenceDTO) {
+        Preference preference = new Preference();
+        preference.setUserId(preferenceDTO.getUserId());
+        preference.setServiceId(preferenceDTO.getServiceId());
+        buildPreference(preferenceDTO, preference);
+
         preference.setCreatedAt(LocalDateTime.now());
         preference.setUpdatedAt(LocalDateTime.now());
         return preferenceRepository.save(preference);
     }
 
-    public Preference updatePreference(UUID id, Preference updatedPreference) {
-        Optional<Preference> optionalPreference = preferenceRepository.findById(id);
-        if (optionalPreference.isPresent()) {
-            Preference preference = optionalPreference.get();
-            preference.setPaymentPreference(updatedPreference.getPaymentPreference());
-            preference.setLocation(updatedPreference.getLocation());
-            preference.setSameLocation(updatedPreference.getSameLocation());
-            preference.setSchedulingPolicy(updatedPreference.getSchedulingPolicy());
-//            preference.setDayOfWeek(updatedPreference.getDayOfWeek());
-//            preference.setFromTime(updatedPreference.getFromTime());
-//            preference.setToTime(updatedPreference.getToTime());
-            preference.setUpdatedAt(LocalDateTime.now());
-            return preferenceRepository.save(preference);
-        } else {
-            throw new IllegalArgumentException("Preference with ID " + id + " not found.");
-        }
+    private void buildPreference(PreferenceDTO preferenceDTO, Preference preference) {
+        preference.setPaymentPreference(preferenceDTO.getPaymentPreference());
+        preference.setLocation(preferenceDTO.getLocation());
+        preference.setSameLocation(preferenceDTO.getSameLocation());
+        preference.setSchedulingPolicy(preferenceDTO.getSchedulingPolicy());
+        List<BookingDay> bookingDays = preferenceDTO.getBookingDays().stream()
+                .map(this::mapToBookingDay)
+                .collect(Collectors.toList());
+
+        preference.setBookingDays(bookingDays);
+    }
+
+    private BookingDay mapToBookingDay(BookingDayDTO dto) {
+        BookingDay bookingDay = new BookingDay();
+        bookingDay.setDayOfWeek(dto.getDayOfWeek());
+        bookingDay.setFromTime(dto.getStartTime());
+        bookingDay.setToTime(dto.getEndTime());
+        return bookingDay;
+    }
+
+    public Preference updatePreference(UUID id, PreferenceDTO updatedPreferenceDTO) {
+        Preference preference = preferenceRepository.findById(id)
+                .orElseThrow(() -> new PreferenceNotFoundException(id));
+        buildPreference(updatedPreferenceDTO, preference);
+
+        preference.setUpdatedAt(LocalDateTime.now());
+
+        return preferenceRepository.save(preference);
     }
 
     public Optional<Preference> getPreferenceById(UUID id) {
-        return preferenceRepository.findById(id);
+        return preferenceRepository.findById(id)
+                .or(() -> { throw new PreferenceNotFoundException(id); });
     }
 
     public List<Preference> getAllPreferences() {
@@ -50,6 +73,11 @@ public class PreferenceService {
     }
 
     public void deletePreference(UUID id) {
+        if (!preferenceRepository.existsById(id)) {
+            throw new PreferenceNotFoundException(id);
+        }
         preferenceRepository.deleteById(id);
     }
+
+
 }
